requirements.txt
pyexiftool
Pillow